Substantially improve naming in auto-generated code.
Before we used ugly-but-always-unique names.  With the new nameMapper,
we try hard to generate nice short names.  If that doesn't work, we fall
back to ugly-but-always-unique names.

In addition, the nameMapper correctly handles much more complicated types
such as chans (they have spaces in the name), inline structs (that
have non-alphanumeric characters such as braces and brackets in them),
and slices.