Logging: Add Quiet flag, NoLog helper, log RequestURI, improve error msg.
The Quiet flag can be set on a LogEntry to indicate that it shouldn't be
logged.  This must be respected by the WriteLog function, and the default
one does.  However, this still allow the WriteLog function to be a nice
hook for tracking overall server statistics.

Add the NoLog helper function that can operate as middleware to suppress
log for certain routes.

Fix logging of the request path.  We now use .RequestURI instead of
.URL.Path since the URL may be modified during the middleware processing,
but the request URI is supposed to remain unchanged.

Lastly, the default string for sandwich.Error will include the ClientMsg
if the LogMsg is empty, which is typical for cases like:
  return sandwich.Error{Code: 404}