BREAKING: API Refactoring
The API changes are motivated by the observation that chains are
basically assembled functions. The intended usage is that you make a
chain of Funcs with Args that you Run. When you run it, it will call
functions specified by the .Then and .Defer calls and provide those
functions the most recent value(s) from the args, .Then return values,
or .Set calls.

Major changes include:
- Rename `Chain`     -> `Func`
- Rename `Reserve`   -> `Arg`
- Rename `Provide`   -> `Set`
- Rename `ProvideAs` -> `SetAs`
- Rename `With`      -> `Then`

These renames occurred in both the chain subpackage and the middleware
interfaces and implementations.

This is in service of working towards a 1.0 stable release.