BREAKING: Make the default error handle panic
The default error handle for the chain package should panic. Only when
it's used in an http middleware setting is it reasonable to do some
alternate default handling, but the middleware package itself should
specify the default error handling.

Just printing an error is extremely dangerous for the default behavior.