BREAKING: Major API refactor: Router replaced Middleware
The lack of a native routing concept in sandwich made it extremely
difficult to use in practice. Adding the routing API was possible but
unwieldy, and had to be added for each router framework -- httprouter
and martini were existing integrations, but httprouter is extremely
limited and martini is old and unsupported.

Newer frameworks like chi are difficult to integrate with because of the
extensive wrapper-style middleware functions that include a `next()`
function. The `next` functionality is difficult to reconcile with the
sandwich philosophy.

Instead, I built a new routing mechanism that allows fairly simple
pat-style routing along with sub-routers. In particular, since speed was
never the main driver behind sandwich, the routing implementation is
tree-based and should be reasonably fast but it doesn't matter too much
since the call chaining is the slowest part of the routing. Even still,
for any meaningful server interactions the routing isn't going to be the
bottleneck, so the convenience still reigns supreme.